@@ -2,7 +2,12 @@
 
 source 'http://rubygems.org'
 
-gem "rails", "3.0.7"
+gem "rails", "3.1.0.rc1"
+# gem "rails", :git => "git://github.com/rails/rails.git"
+
+gem "sass"
+gem "coffee-script"
+gem "jquery-rails"
 
 gem 'unicorn'
 
@@ -20,11 +25,11 @@ gem 'gravatarify', "~> 2.2.2"
 gem 'dynamic_form'
 
 group :test, :development do
-  gem 'capybara', '>=1.0.0.beta1'
+  gem 'capybara', :git => 'git://github.com/jnicklas/capybara.git'
   # gem 'steak'
   gem "database_cleaner"
-  gem 'rspec', ">= 2.6.0.rc6"
-  gem 'rspec-rails', ">= 2.6.0.rc6"
+  gem 'rspec', :git => 'https://github.com/rspec/rspec.git'
+  gem 'rspec-rails', :git => "https://github.com/rspec/rspec-rails.git"
   gem "annotate-models", :require => false
   # gem "remarkable_rails" # does not work yet
   gem "remarkable_activerecord", ">= 4.0.0.alpha4"
@@ -39,5 +44,5 @@ group :test, :development do
   gem 'guard-livereload'
   gem 'growl'
   gem 'hirb'
-  gem 'bullet' # for finding n+1
+#  gem 'bullet' # for finding n+1
 end
