@@ -1,10 +1,8 @@
 # Be sure to restart your server when you modify this file.
 
-Rails.application.config.session_store :cookie_store, {
-  :key => '_echowaves2_session',
-}
+Echowaves::Application.config.session_store :cookie_store, key: '_echowaves_session'
 
 # Use the database for sessions instead of the cookie-based default,
 # which shouldn't be used to store highly confidential information
-# (create the session table with "rake db:sessions:create")
-# Rails.application.config.session_store :active_record_store
+# (create the session table with "rails generate session_migration")
+# Echowaves::Application.config.session_store :active_record_store
