@@ -22,15 +22,15 @@ Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each {|f| require f}
 
 require 'remarkable/active_record'
 
-Rspec.configure do |config|
-  
+RSpec.configure do |config|
+
   config.before(:suite) do
     DatabaseCleaner.strategy = :truncation
   end
   # Remove this line if you don't want Rspec's should and should_not
   # methods or matchers
   require 'rspec/expectations'
-  
+
   # config.include Rspec::Matchers
 
   # == Mock Framework
